requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.0.9
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.2.0
billiard==3.6.4.0
celery==5.2.2
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.0
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.13.1
frozenlist==1.2.0
gunicorn==20.1.0
idna==3.3
kombu==5.2.2
multidict==5.2.0
prompt-toolkit==3.0.24
psycopg2-binary==2.8.6
pytz==2021.3
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
yarl==1.7.2
